pyproject.toml
[project]
name = "desktop-automation"
version = "0.1.0"
description = "Vision-based desktop automation with two-stage VLM grounding"
requires-python = ">=3.11"
dependencies = [
    "google-genai",
    "pyautogui",
    "pillow",
    "requests",
    "pywin32",
    "opencv-python>=4.13.0.92",
    "numpy>=2.4.4",
]

[project.scripts]
automate = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []
